pyproject.toml
[tool.poetry]
name = "chromadb-start"
version = "0.1.0"
description = ""
authors = ["Francisco Javier Martínez Páez <fjmpaez@izertis.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
chromadb = "^0.5.12"
openai = "^1.51.2"
python-dotenv = "^1.0.1"
langchain = "^0.1.0"
langchain-openai = "^0.1.0"

[tool.poetry.dev-dependencies]
pypdf = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
